requirements.txt
amqp==5.1.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.3.0
bs4==0.0.1
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
Deprecated==1.2.13
fake-useragent==0.1.11
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lxml==4.8.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
parse==1.19.0
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
py==1.11.0
pyee==8.2.2
pyparsing==3.0.8
pyppeteer==1.0.2
pyquery==1.4.3
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
requests==2.27.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tqdm==4.64.0
typing-extensions==4.2.0
ultimate-sitemap-parser==0.5
urllib3==1.26.9
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
